Problem: Invite protocol is untested and broken
Solution:
 - Test it, fail and comment out the test (invite_test.exs)
 - Then test all the cryptography that is related to credentials,
   including formatting (format convergence between proofless / idless
   maps, complete maps, and Ecto schema instances)
 - Fix all the bugs we found there
 - Uncomment invite_test.exs and it went green!

Notable ideas, problems and cues for future work:

Since tests are really quick, we can't rely on second-resolved
timestamps to yield different signatures. This is why we needed to add a
random wait into the issuance date.

Moving forward, we should probably add a PgSQL-compatible nanosecond
noise in some or all of our credentials.

As well as that, we should also reuse already-inserted proofs and
subjects but I don't see lack of these features of business logic as a
huge problem at the moment.

That's about it, that's the commit. Enjoy!