Problem: assets are old and incompat. w/ tailwind
Soluition:

 - Tweak getStuff until we found compatible set of versions
 - Add postcss configuration
 - Add tailwind
 - Remove SASS entirely

Here's the explanation of what SASS, PostCSS and TailwindCSS are:

    - SASS is a css like language that compile to CSS. It used to be big
      because CSS was lacking a ton of stuff that people wanted to have.

    - Post CSS is a CSS processor. It takes CSS and all kind of "macros"
      you can define in it with plugins and transform it into the final CSS.
      It is more or less how most of "recent" CSS framework do things, as it
      is pluggable and does not force a language at the entry, just enhance
      css with macros

    - Tailwind use PostCSS for a lot of its functionalities

Now, in the past, phoenix used SASS for its "base" generated code, as it
was felt it was nicer to use and more or less the "Standard" especially
coming from ruby.
SASS has been a pain in the build system side of things, and CSS have also
progressed a lot.

by di4na, as seen on elixir-lang slack.