Problem: No way to run one-shot actions against backend
Solution:

In order to run wdio suites, it would be cool to be able to run one-shot tests against the backend in question.

We implemented it by making sure we have a production build built in Dockerfile. We removed the "ENDPOINT" directive, because, really, it's fairly useless.

Then we used netcat (which is included in today's version of sneakium) in order to have a simple single-message passing protocol between containers so that they know when to do certain actions. This is a huge change to docker-compose.yaml.

We changed run.ts to allow for other react-script verbs (not only start) and reformatted it.