Problem: line counting is too specific
Solution:

To imrove readility and remove potential bugs on other systems, we have spread the "file exists" check and "line counting" check apart.

In the line counting check, we now check if the amount of found lines is greater than zero rather than just checking if it's one or two.