Problem: String is meaningless in Ureus, no verifyMap
Solution:

Here we demonstrate that crypto makes sense overall. We demonstrate that main keys can be used to derive an SKP and sign something.
Not only that, but we also demonstrate that we can take any canonicalisable map (see Cannable type in crypto.ts) and it can be cryptographically signed, creating verifiable credential.
In the meantime, we fix some bugs and we're happy.